Split resolver collection test modules
